Simplify offset handling
We now use int64s everywhere a value refers to a disk offset.
We store the Header.block_size (which must be a power of 2) as a
  power of 2 for easy shifting.
We now prefer to calculate sector addresses rather than byte addresses,
  since everything is naturally sector-aligned.

+++ lib/s.ml
@@ -264,8 +264,8 @@ module type IO = sig
 
   val exists: string -> bool t
   val openfile: string -> fd t
-  val really_read: fd -> int -> int -> Cstruct.t t
-  val really_write: fd -> int -> Cstruct.t -> unit t
+  val really_read: fd -> int64 -> int -> Cstruct.t t
+  val really_write: fd -> int64 -> Cstruct.t -> unit t
 end
